src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "codecarbon>=2.8.4",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "ruff>=0.11.3",
    "python-barcode>=0.15.1",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.3",
]
